+++ Smart-Inventory-Prediction-System/src/main/java/com/Smart/Inventory/Prediction/System/model/SaleItem.java
@@ -3,6 +3,8 @@ package com.Smart.Inventory.Prediction.System.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name="saleItems")
@@ -13,8 +15,8 @@ public class SaleItem {
     private Long id;
 
     private Long quantity;
-    private Double unitPrice;
-    private Double subTotal;
+    private BigDecimal unitPrice;
+    private BigDecimal subTotal;
 
 
     @ManyToOne
